useCommonTextInputProps hook to unify function to get characters left text
